Moved all User related queries into one parentHook
